refactor: organize service extensions and add ExcludeFromCodeCoverage attribute
