+++ src/AStar.Dev.OneDrive.Sync.Client/Accounts/AccountManagementView.axaml.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using Avalonia.Controls;
 
 namespace AStar.Dev.OneDrive.Sync.Client.Accounts;
@@ -5,6 +6,7 @@ namespace AStar.Dev.OneDrive.Sync.Client.Accounts;
 /// <summary>
 ///     View for managing OneDrive accounts.
 /// </summary>
+[ExcludeFromCodeCoverage]
 public partial class AccountManagementView : UserControl
 {
     /// <summary>
