+++ src/AStar.Dev.OneDrive.Sync.Client/Settings/SettingsWindow.axaml.cs
@@ -1,8 +1,10 @@
+using System.Diagnostics.CodeAnalysis;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 
 namespace AStar.Dev.OneDrive.Sync.Client.Settings;
 
+[ExcludeFromCodeCoverage]
 public partial class SettingsWindow : Window
 {
     public SettingsWindow() => InitializeComponent();
