+++ src/AStar.Dev.OneDrive.Sync.Client/Syncronisation/SyncProgressView.axaml.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using Avalonia.Controls;
 
 namespace AStar.Dev.OneDrive.Sync.Client.Syncronisation;
@@ -5,6 +6,7 @@ namespace AStar.Dev.OneDrive.Sync.Client.Syncronisation;
 /// <summary>
 ///     Code-behind for the sync progress view.
 /// </summary>
+[ExcludeFromCodeCoverage]
 public partial class SyncProgressView : UserControl
 {
     /// <summary>
