+++ src/AStar.Dev.OneDrive.Sync.Client/SyncronisationConflicts/ConflictResolutionView.axaml.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using Avalonia.Controls;
 
 namespace AStar.Dev.OneDrive.Sync.Client.SyncronisationConflicts;
@@ -5,6 +6,7 @@ namespace AStar.Dev.OneDrive.Sync.Client.SyncronisationConflicts;
 /// <summary>
 ///     Code-behind for the conflict resolution view.
 /// </summary>
+[ExcludeFromCodeCoverage]
 public partial class ConflictResolutionView : UserControl
 {
     /// <summary>
